:bug: Move onActivityResult logic to projectOverviewFragment
