+++ app/src/main/java/com/kyhsgeekcode/disassembler/IDrawerManager.kt
@@ -1,0 +1,5 @@
+package com.kyhsgeekcode.disassembler
+
+interface IDrawerManager {
+    fun notifyDataSetChanged()
+}
